refactor: Moves database extensions to correct file.
